requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
bip-utils==2.9.3
bip32utils==0.3.post4
bitarray==2.9.2
cbor2==5.6.4
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
ckzg==2.0.0
coincurve==20.0.0
crcmod==1.7
cytoolz==0.12.3
Django==4.2.13
djangorestframework==3.15.1
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
filelock==3.15.4
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.7
mnemonic==0.21
multidict==6.0.5
parsimonious==0.10.0
pillow==10.4.0
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pypng==0.20220715.0
python-dotenv==1.0.1
pyunormalize==15.1.0
qrcode==7.4.2
regex==2024.7.24
requests==2.32.2
rlp==4.0.1
six==1.16.0
sqlparse==0.5.0
toolz==0.12.1
types-requests==2.32.0.20240712
typing_extensions==4.11.0
urllib3==2.2.1
web3==7.2.0
websockets==13.0.1
whitenoise==6.6.0
yarl==1.9.7
